requirements.txt
streamlit==1.24.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1
google-generativeai==0.1.0
